refactor(api): extract EMPTY_BLANK_AND_NULL constant from test-helper.js
