+++ api/tests/certification/session-management/unit/infrastructure/serializers/session-serializer_test.js
@@ -1,7 +1,8 @@
 import { SessionManagement } from '../../../../../../src/certification/session-management/domain/models/SessionManagement.js';
 import * as serializer from '../../../../../../src/certification/session-management/infrastructure/serializers/session-serializer.js';
 import { SESSION_STATUSES } from '../../../../../../src/certification/shared/domain/constants.js';
-import { EMPTY_BLANK_AND_NULL, expect } from '../../../../../test-helper.js';
+import { expect } from '../../../../../test-helper.js';
+import { EMPTY_BLANK_AND_NULL } from '../../../../../tooling/test-utils/constants.js';
 
 describe('Unit | Certification | session-management | Serializer | session-serializer', function () {
   describe('#serialize()', function () {
