+++ api/tests/tooling/test-utils/constants.js
@@ -1,0 +1,5 @@
+/*
+ * Contants used in tests
+ */
+
+export const EMPTY_BLANK_AND_NULL = ['', '\t \n', null];
